cryptage de l'email

+++ backend/controllers/sauce.js
@@ -59,6 +59,10 @@ exports.modifySauce = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
       .then(sauce => {
+        if (sauce.userId !== req.user) {  // on compare l'id de l'auteur de la sauce et l'id de l'auteur de la requête
+          res.status(401).json({message: "action non autorisée"});  // si ce ne sont pas les mêmes id = code 401: unauthorized.
+          return sauce;
+        }
         const filename = sauce.imageUrl.split('/images/')[1];
         fs.unlink(`images/${filename}`, () => {
           Sauce.deleteOne({ _id: req.params.id })
